experiments/hpo/searl/SEARL/requirements.txt
cffi==1.14.5
cloudpickle==1.3.0
Cython==0.29.22
fasteners==0.16
fastrand==1.3.0
future==0.18.2
glfw==2.1.0
gym==0.17.1
imageio==2.9.0
mujoco-py==2.0.2.9
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pyaml==20.4.0
pycparser==2.20
pyglet==1.5.0
PyYAML==5.4.1
scipy==1.6.1
six==1.15.0
torch==1.6.0
